+++ diive/preprocessing/corrections/__init__.py
@@ -9,8 +9,10 @@ Part of the diive library: https://github.com/holukas/diive
 
 from diive.preprocessing.corrections.offsetcorrection import (
     MeasurementOffsetFromReplicate,
+    NighttimeZeroOffsetResult,
+    nighttime_zero_offset_diagnostics,
     remove_relativehumidity_offset,
-    remove_radiation_zero_offset,
+    remove_nighttime_zero_offset,
     WindDirOffset,
 )
 from diive.preprocessing.corrections.setto import (
@@ -22,8 +24,10 @@ from diive.preprocessing.corrections.apply import apply_corrections
 
 __all__ = [
     'MeasurementOffsetFromReplicate',
+    'NighttimeZeroOffsetResult',
+    'nighttime_zero_offset_diagnostics',
     'remove_relativehumidity_offset',
-    'remove_radiation_zero_offset',
+    'remove_nighttime_zero_offset',
     'WindDirOffset',
     'set_exact_values_to_missing',
     'setto_threshold',